Remove ADAL support and use MSAL only
