+++ HISTORY.rst
@@ -5,8 +5,8 @@ Release History
 
 0.0.53 (2023-04-11)
 +++++++++++++++++++
+* Add MSAL support. Remove ADAL support
 * Suppress deprecation warning when detecting pyopenssl existence.
-* Add MSAL support
 
 0.0.52 (2020-11-25)
 +++++++++++++++++++
