+++ azure/datalake/store/retry.py
@@ -74,55 +74,3 @@ class ExponentialRetryPolicy(RetryPolicy):
     def __backoff(self):
         time.sleep(self.exponential_retry_interval)
         self.exponential_retry_interval *= self.exponential_factor
-
-
-def retry_decorator_for_auth(retry_policy = None):
-    import adal
-    from requests import HTTPError
-    if retry_policy is None:
-        retry_policy = ExponentialRetryPolicy(max_retries=2)
-
-    def deco_retry(func):
-        @wraps(func)
-        def f_retry(*args, **kwargs):
-            retry_count = -1
-            while True:
-                last_exception = None
-                retry_count += 1
-                try:
-                    out = func(*args, **kwargs)
-                except (adal.adal_error.AdalError, HTTPError) as e:
-                    # ADAL error corresponds to everything but 429, which bubbles up HTTP error.
-                    last_exception = e
-                    logger.exception("Retry count " + str(retry_count) + "Exception :" + str(last_exception))
-                    # We don't want to stop retry for any error in parsing the exception. This is a GET operation.
-                    try:
-                        if hasattr(last_exception, 'error_response'):  # ADAL exception
-                            response = response_from_adal_exception(last_exception)
-                        if hasattr(last_exception, 'response'):  # HTTP exception i.e 429
-                            response = last_exception.response
-                    except:
-                        pass
-                request_successful = last_exception is None or (response is not None and response.status_code == 401)  # 401 = Invalid credentials
-                if request_successful or not retry_policy.should_retry(response, last_exception, retry_count):
-                    break
-            if last_exception is not None:
-                raise last_exception
-            return out
-        return f_retry
-
-    return deco_retry
-
-
-def response_from_adal_exception(e):
-    import re
-    from collections import namedtuple
-
-    http_code = re.search(r"http error: (\d+)", str(e))
-    if http_code is not None:  # Add status_code to response object for use in should_retry
-        status_code = [int(http_code.group(1))]
-        Response = namedtuple("Response", ['status_code'])
-        response = Response(
-            *status_code)  # Construct response object with adal exception response and http code
-    return response
-
